feat: add main VodkaApplication class fix: change Long to Integer's in all id fields in entity classes

+++ order-accepted/src/main/java/com/rszumlas/orderaccepted/OrderAcceptedId.java
@@ -15,9 +15,9 @@ import java.io.Serializable;
 public class OrderAcceptedId implements Serializable {
 
     @Column(name = "order_id")
-    private Long order_id;
+    private Integer order_id;
 
     @Column(name = "employee_id")
-    private Long employee_id;
+    private Integer employee_id;
 
 }
